fixed unsigned issue.  Bonus slots fixed.  Back button lobby controls added - Backspace to return to lobby from game over

+++ Dev/StaticSheepNeoBuildSystem/game/SheepEngine/Components/slotmachine/slotmachine.cpp
@@ -254,7 +254,7 @@ namespace Framework
                 : nullptr;
 
               selectionCB.ForceBind(context);
-              selectionCB(i, &(slot->land));
+              selectionCB((int)i, &(slot->land));
             }
               
             else
